Refactor column config on Event Actions page to fix lint error line numbers exceeded
